requirements.txt
# force numpy <2 so C‐extensions (pyarrow, faiss‐cpu, etc.) keep working
numpy<2

# core libs
pandas>=1.3
sentence-transformers>=2.2
torch>=1.12

# FAISS for ANN search (CPU build)
faiss-cpu>=1.10.0

# lexical retrieval
rank_bm25>=0.2.2

# evaluation metrics
scikit-learn>=1.0

# web app
streamlit>=1.10

# pyarrow must match numpy<2
pyarrow>=9.0.0,<13.0.0
